+++ src/main/java/com/example/bank/components/CreditEditor.java
@@ -2,39 +2,31 @@ package com.example.bank.components;
 
 import com.example.bank.domain.Credit;
 import com.example.bank.repo.CreditRepo;
+import com.example.bank.services.CreditOfferService;
 import com.vaadin.flow.component.Key;
 import com.vaadin.flow.component.KeyNotifier;
 import com.vaadin.flow.component.button.Button;
 import com.vaadin.flow.component.icon.VaadinIcon;
+import com.vaadin.flow.component.notification.Notification;
 import com.vaadin.flow.component.orderedlayout.HorizontalLayout;
 import com.vaadin.flow.component.orderedlayout.VerticalLayout;
 import com.vaadin.flow.component.textfield.TextField;
 import com.vaadin.flow.data.binder.Binder;
-import com.vaadin.flow.data.converter.StringToDoubleConverter;
-import com.vaadin.flow.data.converter.StringToLongConverter;
+import com.vaadin.flow.data.converter.StringToBigDecimalConverter;
 import com.vaadin.flow.spring.annotation.SpringComponent;
 import com.vaadin.flow.spring.annotation.UIScope;
 import lombok.Setter;
 
-import org.springframework.beans.factory.annotation.Autowired;
+import java.math.BigDecimal;
 
 @SpringComponent
 @UIScope
 public class CreditEditor extends VerticalLayout implements KeyNotifier {
-
     private final CreditRepo creditRepo;
+    private final CreditOfferService creditOfferService;
+    private final TextField name = new TextField("Имя кредита");
+    private final Binder<Credit> binder = new Binder<>(Credit.class);
     private Credit credit;
-    private TextField limit = new TextField("Лимит по кредиту");
-    private TextField interestRate = new TextField("Процентная ставка");
-    private TextField name = new TextField("Имя кредита");
-
-
-    private Button save = new Button("Save", VaadinIcon.CHECK.create());
-    private Button cancel = new Button("Cancel");
-    private Button delete = new Button("Delete", VaadinIcon.TRASH.create());
-    private HorizontalLayout actions = new HorizontalLayout(save, cancel, delete);
-    private HorizontalLayout fields = new HorizontalLayout(limit, interestRate, name);
-    private Binder<Credit> binder = new Binder<>(Credit.class);
     @Setter
     private CreditEditor.ChangeHandler changeHandler;
 
@@ -42,18 +34,30 @@ public class CreditEditor extends VerticalLayout implements KeyNotifier {
         void onChange();
     }
 
-
-    @Autowired
-    public CreditEditor(CreditRepo creditRepo) {
+    public CreditEditor(CreditRepo creditRepo, CreditOfferService creditOfferService) {
         this.creditRepo = creditRepo;
+        this.creditOfferService = creditOfferService;
+        Button save = new Button("Save", VaadinIcon.CHECK.create());
+        Button cancel = new Button("Cancel");
+        Button delete = new Button("Delete", VaadinIcon.TRASH.create());
+        HorizontalLayout actions = new HorizontalLayout(save, cancel, delete);
+        TextField interestRate1 = new TextField("Процентная ставка");
+        TextField limit1 = new TextField("Лимит по кредиту");
+        HorizontalLayout fields = new HorizontalLayout(limit1, interestRate1, name);
         add(fields, actions);
 
-        binder.forField(limit)
-                .withConverter(new StringToLongConverter("Only digits allowed"))
+        binder.forField(limit1)
+                .withConverter(new StringToBigDecimalConverter("Only digits allowed"))
+                .withValidator(
+                        limit -> limit.compareTo(BigDecimal.ZERO) > 0,
+                        "Limit rate must be greater than zero")
                 .bind(Credit::getLimit, Credit::setLimit);
 
-        binder.forField(interestRate)
-                .withConverter(new StringToDoubleConverter("Only digits allowed"))
+        binder.forField(interestRate1)
+                .withConverter(new StringToBigDecimalConverter("Only digits allowed"))
+                .withValidator(
+                        interestRate -> interestRate.compareTo(BigDecimal.ZERO) > 0,
+                        "Interest rate must be greater than zero")
                 .bind(Credit::getInterestRate, Credit::setInterestRate);
 
         binder.forField(name)
@@ -86,6 +90,12 @@ public class CreditEditor extends VerticalLayout implements KeyNotifier {
     }
 
     private void delete() {
+        if (!creditOfferService.findAllByCredit(credit).isEmpty()) {
+            Notification notification = new Notification(
+                    "Этот кредит используется в кредитных предложениях!", 5000);
+            notification.open();
+            return;
+        }
         creditRepo.delete(credit);
         changeHandler.onChange();
     }
